Moved BiomeProviderRegisterEvent to correct package, Reference2IntOpenHashMap proper usage
