+++ station-worldgen-api-v0/src/main/java/net/modificationstation/stationapi/impl/worldgen/WorldgenListener.java
@@ -4,8 +4,8 @@ import net.mine_diver.unsafeevents.listener.EventListener;
 import net.mine_diver.unsafeevents.listener.ListenerPriority;
 import net.modificationstation.stationapi.api.StationAPI;
 import net.modificationstation.stationapi.api.event.level.biome.BiomeRegisterEvent;
+import net.modificationstation.stationapi.api.event.worldgen.biome.BiomeProviderRegisterEvent;
 import net.modificationstation.stationapi.api.worldgen.BiomeAPI;
-import net.modificationstation.stationapi.api.event.BiomeProviderRegisterEvent;
 
 public class WorldgenListener {
 	private boolean initiated;
