+++ station-worldgen-api-v0/src/main/java/net/modificationstation/stationapi/mixin/worldgen/MixinBiomeSource.java
@@ -10,7 +10,6 @@ import net.minecraft.level.gen.BiomeSource;
 import net.modificationstation.stationapi.api.worldgen.BiomeAPI;
 import net.modificationstation.stationapi.api.worldgen.biome.BiomeProvider;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
